+++ examples/example_host.h
@@ -9,6 +9,8 @@
 extern "C" {
 #endif
 
+const struct evmc_host_interface* example_host_get_interface();
+
 struct evmc_host_context* example_host_create_context(struct evmc_tx_context tx_context);
 
 void example_host_destroy_context(struct evmc_host_context* context);